+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 *   New Features
     *   Add option to switch to the HLS video stream when playing a downloaded episode
         ([#5620](https://github.com/Automattic/pocket-casts-android/pull/5620))
+*   Bug Fixes
+    *   Fix playback repeatedly pausing on wireless Android Auto
+        ([#5535](https://github.com/Automattic/pocket-casts-android/pull/5535))
 
 8.17
 -----